lots of refinements and bug fixes

+++ src/server/hosts/service.rs
@@ -10,7 +10,7 @@ use crate::server::{
         types::{
             base::{Host, HostBase}
         }
-    }, interfaces::types::base::{Interface, InterfaceBase}, services::{service::ServiceService, types::base::{Service}}, subnets::service::SubnetService
+    }, services::{service::ServiceService}, subnets::service::SubnetService
 };
 
 pub struct HostService {
@@ -73,64 +73,13 @@ impl HostService {
     }
 
     pub async fn consolidate_hosts(&self, mut destination_host: Host, other_host: Host) -> Result<Host> {
-        let mut new_interfaces: Vec<Interface> = Vec::new();
         let other_host_services = self.service_service.get_services_for_host(&other_host.id).await?;
 
-        let other_host_services_updates: Vec<Service> = other_host_services.into_iter().map(|mut s| {
-                                    
-            // Update bindings - check for subnet compatibility, not interface ID matching
-            s.base.interface_bindings = s.base.interface_bindings.iter().filter_map(|binding_id| {
-                
-                // Get the original interface from the other host
-                if let Some(origin_interface) = other_host.get_interface(binding_id) {
-                    
-                    // Check if destination host already has an interface on the same subnet
-                    if let Some(dest_interface) = destination_host.base.interfaces.iter()
-                        .find(|dest_iface| dest_iface.base.subnet_id == origin_interface.base.subnet_id) {
-                        
-                        // Reuse existing destination interface
-                        Some(dest_interface.id)
-                        
-                    } else {
-                        
-                        // Check if we've already decided to create this interface
-                        if let Some(existing_new) = new_interfaces.iter()
-                            .find(|new_iface| new_iface.base.subnet_id == origin_interface.base.subnet_id) {
-                                                    
-                            Some(existing_new.id)
-                            
-                        } else {
-                            
-                            // Create new interface based on origin, but with new ID
-                            let new_interface = Interface::new(InterfaceBase {
-                                name: origin_interface.base.name.clone(),
-                                subnet_id: origin_interface.base.subnet_id,
-                                ip_address: origin_interface.base.ip_address,
-                                mac_address: None,
-                            });
-                            
-                            let interface_id = new_interface.id;
-                            new_interfaces.push(new_interface);
-                            Some(interface_id)
-                        }
-                    }
-                } else {
-                    tracing::warn!("Could not find interface {} on origin host", binding_id);
-                    None
-                }
-            })
-            .collect();
+        destination_host.base.interfaces = [destination_host.base.interfaces, other_host.base.interfaces].concat();
+        destination_host.base.services = [destination_host.base.services, other_host_services.iter().map(|s| s.id).collect()].concat();
 
+        let service_update_futures = other_host_services.into_iter().map(|mut s| {
             s.base.host_id = destination_host.id;
-
-            s
-        })
-        .collect();
-
-        destination_host.base.interfaces = [destination_host.base.interfaces, new_interfaces].concat();
-        destination_host.base.services = [destination_host.base.services, other_host_services_updates.iter().map(|s| s.id).collect()].concat();
-
-        let service_update_futures = other_host_services_updates.into_iter().map(|s| {
             self.service_service.update_service(s)
         });
 
@@ -138,7 +87,9 @@ impl HostService {
 
         let updated_host = self.update_host(destination_host).await?;
 
-        self.delete_host(&other_host.id).await?;
+        tracing::info!("{:?}", updated_host);
+
+        self.delete_host(&other_host.id, true).await?;
 
         Ok(updated_host)
     }
@@ -164,15 +115,17 @@ impl HostService {
         Ok(())
     }
 
-    pub async fn delete_host(&self, id: &Uuid) -> Result<()> {
+    pub async fn delete_host(&self, id: &Uuid, ignore_services: bool) -> Result<()> {
 
         let host = self.get_host(id).await?.ok_or_else(|| anyhow::anyhow!("Host {} not found", id))?;
 
-        let service_deletion_futures = host.base.services.iter().map(|s| {
-            self.service_service.delete_service(s, false)
-        });
+        if !ignore_services {
+            let service_deletion_futures = host.base.services.iter().map(|s| {
+                self.service_service.delete_service(s, false)
+            });
 
-        try_join_all(service_deletion_futures).await?;
+            try_join_all(service_deletion_futures).await?;
+        }
         
         self.update_subnet_host_relationships(&host, true).await?;
